cloud_function/requirements.txt
google-cloud-language==2.13.0
google-cloud-bigquery==3.21.0
functions-framework==3.8.1
